requirements.txt
asgiref==3.5.2
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
Cython==0.29.32
defusedxml==0.7.1
Django==4.1.3
django-debug-toolbar==3.7.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
drf-nested-routers==0.93.4
drf-writable-nested==0.7.0
exceptiongroup==1.1.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
locust==2.14.2
lxml==4.9.2
MarkupSafe==2.1.1
model-bakery==1.9.0
msgpack==1.0.4
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
platformdirs==2.6.2
pluggy==1.0.0
psutil==5.9.4
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
pytest-env==0.8.1
pytest-watch==4.2.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rest-framework-generic-relations==2.1.0
roundrobin==0.0.4
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
watchdog==2.2.1
Werkzeug==2.2.2
zeep==4.2.1
zope.event==4.6
zope.interface==5.5.2
